now reestimating split A appropriately this thing kind of works... kind of.

+++ cpphmm/src/HmmTypes.h
@@ -6,7 +6,7 @@
 #include <stdint.h>
 #include <iostream>
 
-typedef double HmmFloat_t;
+typedef float HmmFloat_t;
 typedef std::vector<HmmFloat_t> HmmDataVec_t;
 typedef std::vector<HmmDataVec_t> HmmDataMatrix_t;
 typedef std::vector<HmmDataMatrix_t> Hmm3DMatrix_t;
